refactor: renamed the service directory
