+++ frontend/Barcode/src/assets/services/Barcode.js
